+++ packages/client/src/observable/internal/ChangedObjects.ts
@@ -16,18 +16,57 @@
 
 import type { ObjectTypeDefinition, Osdk } from "@osdk/api";
 import { MultiMap } from "mnemonist";
+import { DEBUG_ONLY__cacheKeyToString } from "./CacheKey.js";
+import type { ListCacheKey } from "./ListQuery.js";
 
-export interface ChangedObjects {
+export interface Changes {
   modifiedObjects: MultiMap<string, Osdk.Instance<ObjectTypeDefinition>>;
   addedObjects: MultiMap<string, Osdk.Instance<ObjectTypeDefinition>>;
+  addedLists: Set<ListCacheKey>;
+  modifiedLists: Set<ListCacheKey>;
 }
 
-export function createChangedObjects(): ChangedObjects {
+export function createChangedObjects(): Changes {
   return {
     modifiedObjects: new MultiMap<
       string,
       Osdk.Instance<ObjectTypeDefinition>
     >(),
     addedObjects: new MultiMap<string, Osdk.Instance<ObjectTypeDefinition>>(),
+    addedLists: new Set<ListCacheKey>(),
+    modifiedLists: new Set<ListCacheKey>(),
   };
+}
+
+export function DEBUG_ONLY__changesToString(changes: Changes): string {
+  if (process.env.NODE_ENV !== "production") {
+    return JSON.stringify(
+      {
+        modifiedObjects: multimapHelper(changes.modifiedObjects),
+        addedObjects: multimapHelper(changes.addedObjects),
+        addedLists: listHelper(changes.addedLists),
+        modifiedLists: listHelper(changes.modifiedLists),
+      },
+      null,
+      2,
+    );
+  } else {
+    throw new Error("not implemented");
+  }
+}
+
+function listHelper(set: Set<ListCacheKey>) {
+  return Array.from(set).map(DEBUG_ONLY__cacheKeyToString);
+}
+
+function multimapHelper(
+  multimap: MultiMap<string, Osdk.Instance<ObjectTypeDefinition>>,
+) {
+  return Object.fromEntries(
+    Array.from(multimap.associations()).map(
+      ([type, objects]) => {
+        return [type, objects.map(o => o.$primaryKey)];
+      },
+    ),
+  );
 }