+++ packages/client/src/observable/internal/testUtils.ts
@@ -17,45 +17,240 @@
 import type {
   ActionDefinition,
   ActionEditResponse,
+  ObjectOrInterfaceDefinition,
   ObjectSet,
   ObjectTypeDefinition,
   Osdk,
   OsdkBase,
+  PageResult,
+  WhereClause,
 } from "@osdk/api";
+import { Chalk } from "chalk";
 import type { DeferredPromise } from "p-defer";
 import pDefer from "p-defer";
+import invariant from "tiny-invariant";
 import type { Mock } from "vitest";
 import { afterEach, beforeEach, expect, vi, vitest } from "vitest";
 import type { ActionSignatureFromDef } from "../../actions/applyAction.js";
 import type { Client } from "../../Client.js";
+import { additionalContext } from "../../Client.js";
+import type { LogFn, Logger } from "../../Logger.js";
 import type { ListPayload } from "../ListPayload.js";
 import type { ObjectPayload } from "../ObjectPayload.js";
 import type { Status, Unsubscribable } from "../ObservableClient.js";
 import type { Entry } from "./Layer.js";
+
+const chalk = new Chalk(); // new Chalk({ level: 3 });
 
 export interface MockClientHelper {
   client: Mock<Client> & Client;
 
   mockApplyActionOnce: () => DeferredPromise<Partial<ActionEditResponse>>;
 
-  mockFetchOneOnce: <Q extends ObjectTypeDefinition>() => DeferredPromise<
+  mockFetchOneOnce: <Q extends ObjectTypeDefinition>(
+    expectedId?: string | number | boolean,
+  ) => DeferredPromise<
     Partial<
       Osdk.Instance<Q>
     >
   >;
+
+  mockObjectFactory2Once: () => DeferredPromise<
+    Osdk.Instance<ObjectOrInterfaceDefinition, never, any, {}>[]
+  >;
+
+  mockFetchPageOnce: <
+    Q extends ObjectTypeDefinition,
+  >(
+    expected?: {
+      where: WhereClause<any>;
+      orderBy: Record<string, "asc" | "desc" | undefined>;
+    },
+  ) => DeferredPromise<
+    PageResult<Partial<Osdk.Instance<Q>>>
+  >;
+}
+
+function mockLog(...args: any[]) {
+  // eslint-disable-next-line no-console
+  console.log(
+    chalk.yellow("mockClient"),
+    ...args,
+  );
+}
+// interface LogFn {
+//   (obj: unknown, msg?: string, ...args: any[]): void;
+//   (msg: string, ...args: any[]): void;
+// }
+
+export function createTestLogger(
+  bindings: Record<string, any>,
+  options?: { level?: string; msgPrefix?: string },
+): Logger {
+  const colors = {
+    debug: [chalk.cyan, chalk.bgCyan],
+    info: [chalk.green, chalk.bgGreen],
+    trace: [chalk.gray, chalk.bgGray],
+    error: [chalk.red, chalk.bgRed],
+    warn: [chalk.yellow, chalk.bgYellow],
+    fatal: [chalk.redBright, chalk.bgRedBright],
+  } as const;
+  function createLogMethod(
+    name: "debug" | "error" | "info" | "warn" | "fatal" | "trace",
+  ) {
+    return vi.fn<LogFn>(
+      (
+        ...args: [
+          obj: unknown,
+          ...args1: any[],
+        ] | [
+          ...args2: any[],
+        ]
+      ) => {
+        const hasData = typeof args[0] !== "string";
+        const obj: Record<string, unknown> = hasData ? args[0] as any : {};
+        const more: any[] = hasData ? args.slice(1) : args.slice(0);
+
+        // eslint-disable-next-line no-console
+        console.log(
+          `${colors[name][1](name)}${
+            options?.msgPrefix ? " " + colors[name][0](options.msgPrefix) : ""
+          }${obj.methodName ? ` .${chalk.magenta(obj.methodName)}()` : ""}`,
+          ...more,
+        );
+        if (bindings && Object.keys(bindings).length > 0) {
+          // eslint-disable-next-line no-console
+          console.log(bindings);
+        }
+      },
+    ) as LogFn;
+  }
+  return {
+    debug: createLogMethod("debug"),
+    error: createLogMethod("error"),
+    info: createLogMethod("info"),
+    warn: createLogMethod("warn"),
+    fatal: createLogMethod("fatal"),
+    child: vi.fn<Logger["child"]>((theseBindings, theseOptions) =>
+      createTestLogger({
+        ...bindings,
+        ...theseBindings,
+      }, {
+        level: (theseOptions ?? options)?.level,
+        msgPrefix: options?.msgPrefix || theseOptions?.msgPrefix
+          ? `${options?.msgPrefix ? `${options.msgPrefix} ` : ""}${
+            theseOptions?.msgPrefix || ""
+          }`
+          : undefined,
+      })
+    ),
+    trace: createLogMethod("trace"),
+    isLevelEnabled: vi.fn((args) => true),
+  };
 }
 
 export function createClientMockHelper(): MockClientHelper {
   const client = vitest.fn<typeof client>() as unknown as Mock<Client> & Client;
 
+  const logger = createTestLogger({});
+
+  // this is just a fallback for when there is nothing set so we can track whats up
+  client.mockImplementation((...args: any[]) => {
+    const localLogger = logger.child({}, { msgPrefix: "fallback mock client" });
+    try {
+      throw new Error("IN THE FALLBACK MOCK CLIENT IMPLEMENTATION");
+    } catch (e) {
+      localLogger.error("Just for the stack trace", e);
+    }
+
+    let where;
+    const deadObjectSet = {
+      where: (...whereArgs: any[]) => {
+        localLogger.trace("where", whereArgs);
+        where = whereArgs;
+        return deadObjectSet;
+      },
+      fetchPage: (...fetchArgs: any[]) => {
+        localLogger.trace("etchPage", where!, fetchArgs);
+        throw new Error("NO");
+      },
+      fetchOne: (...fetchArgs: any[]) => {
+        localLogger.trace("fetchOne", fetchArgs);
+
+        throw new Error("NO");
+      },
+    };
+    return deadObjectSet;
+  });
+
+  client[additionalContext] = {
+    baseUrl: "http://localhost:8080",
+    ontologyRid: "ri.something",
+    objectFactory: vitest.fn(),
+    objectFactory2: vitest.fn(),
+    ontologyProvider: {
+      getActionDefinition: vitest.fn(),
+      getInterfaceDefinition: vitest.fn(),
+      getObjectDefinition: vitest.fn(),
+      getQueryDefinition: vitest.fn(),
+    },
+    tokenProvider: vitest.fn(),
+    objectSetFactory: vitest.fn(),
+    fetch: vitest.fn(),
+    clientCacheKey: {} as any,
+    requestContext: {},
+    logger,
+  };
+
+  function mockObjectFactory2Once() {
+    const d = pDefer<
+      Osdk.Instance<ObjectOrInterfaceDefinition, never, any, {}>[]
+    >();
+    vi.mocked(client[additionalContext].objectFactory2).mockReturnValueOnce(
+      d.promise,
+    );
+    return d;
+  }
+
+  function mockFetchPageOnce<
+    X extends PageResult<
+      Osdk.Instance<ObjectTypeDefinition>
+    >,
+  >(): DeferredPromise<X> {
+    const d = pDefer<X>();
+
+    const objectSet: ObjectSet<ObjectTypeDefinition> = {
+      fetchPage: async (fetchPageArgs) => {
+        mockLog("fetchPage", fetchPageArgs);
+        const r = await d.promise;
+        return { ...r, $primaryKey: fetchPageArgs };
+      },
+      where: (clause) => {
+        mockLog("where", clause);
+        return objectSet;
+      },
+    } as Pick<
+      ObjectSet<ObjectTypeDefinition>,
+      "fetchPage" | "where"
+    > as ObjectSet<ObjectTypeDefinition>;
+
+    client.mockReturnValueOnce(objectSet);
+    return d;
+  }
+
   function mockFetchOneOnce<
     X extends Partial<OsdkBase<any>>,
-  >(): DeferredPromise<X> {
+  >(expectedId?: string | number | boolean): DeferredPromise<X> {
     const d = pDefer<X>();
 
     client.mockReturnValueOnce(
       {
         fetchOne: async (a) => {
+          mockLog("fetchOne", a);
+          invariant(
+            expectedId === undefined || a === expectedId,
+            "expected id to match",
+          );
           const r = await d.promise;
           return { ...r, $primaryKey: a };
         },
@@ -77,6 +272,8 @@ export function createClientMockHelper(): MockClientHelper {
             type: "edits",
             addedLinks: x.addedLinks ?? [],
             addedObjects: x.addedObjects ?? [],
+            deletedObjects: x.deletedObjects ?? [],
+            deletedLinks: x.deletedLinks ?? [],
             deletedLinksCount: x.deletedLinksCount ?? 0,
             deletedObjectsCount: x.deletedObjectsCount ?? 0,
             editedObjectTypes: x.editedObjectTypes ?? [],
@@ -95,6 +292,8 @@ export function createClientMockHelper(): MockClientHelper {
     client,
     mockApplyActionOnce,
     mockFetchOneOnce,
+    mockObjectFactory2Once,
+    mockFetchPageOnce,
   };
 }
 
@@ -121,12 +320,14 @@ export function createDefer() {
 export function expectSingleListCallAndClear<T extends ObjectTypeDefinition>(
   subFn: Mock<(e: ListPayload | undefined) => void>,
   resolvedList: Osdk.Instance<T>[],
+  payloadOptions: Omit<Partial<ListPayload>, "resolvedList"> = {},
 ): void {
   if (vitest.isFakeTimers()) {
     vitest.runOnlyPendingTimers();
   }
   expect(subFn).toHaveBeenCalledExactlyOnceWith(
-    expect.objectContaining({
+    listPayloadContaining({
+      ...payloadOptions,
       resolvedList,
     }),
   );
@@ -151,9 +352,16 @@ export async function waitForCall(
   subFn: Mock<(e: any) => void>,
   times: number = 1,
 ): Promise<void> {
-  await vi.waitFor(() => {
-    expect(subFn).toHaveBeenCalledTimes(times);
-  });
+  try {
+    await vi.waitFor(() => {
+      expect(subFn).toHaveBeenCalledTimes(times);
+    });
+  } catch (e) {
+    // we don't need the error, it will retrigger on the next line
+    // and that provides better behavior in the vitest vscode
+    // plugin. This places the error in the test itself instead of
+    // only in this file
+  }
   expect(subFn).toHaveBeenCalledTimes(times);
 }