+++ packages/client/src/util/toDataValueQueries.ts
@@ -18,7 +18,10 @@ import type { QueryDataTypeDefinition } from "@osdk/api";
 import { type DataValue } from "@osdk/foundry.ontologies";
 import * as OntologiesV2 from "@osdk/foundry.ontologies";
 import type { MinimalClient } from "../MinimalClientContext.js";
-import { isAttachmentUpload } from "../object/AttachmentUpload.js";
+import {
+  isAttachmentFile,
+  isAttachmentUpload,
+} from "../object/AttachmentUpload.js";
 import { getWireObjectSet, isObjectSet } from "../objectSet/createObjectSet.js";
 import { isOsdkBaseObject } from "./isOsdkBaseObject.js";
 import { isWireObjectSet } from "./WireObjectSet.js";
@@ -39,6 +42,18 @@ export async function toDataValueQueries(
   }
 
   if (Array.isArray(value) && desiredType.multiplicity) {
+    const values = Array.from(value);
+    if (
+      values.some((dataValue) =>
+        isAttachmentUpload(dataValue) || isAttachmentFile(dataValue)
+      )
+    ) {
+      const converted = [];
+      for (const value of values) {
+        converted.push(await toDataValueQueries(value, client, desiredType));
+      }
+      return converted;
+    }
     const promiseArray = Array.from(
       value,
       async (innerValue) =>
@@ -61,7 +76,7 @@ export async function toDataValueQueries(
       }
 
       if (
-        typeof value === "object" && value instanceof Blob && "name" in value
+        isAttachmentFile(value)
       ) {
         const attachment = await OntologiesV2.Attachments.upload(
           client,