+++ packages/generator/src/shared/getObjectTypesFromQueryDataType.ts
@@ -63,6 +63,7 @@ export function getObjectTypesFromQueryDataType(
     case "threeDimensionalAggregation":
     case "timestamp":
     case "twoDimensionalAggregation":
+    case "entrySet":
     case "unsupported":
       /* complete no-op */
       return;