+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @osdk/client.test.ontology
 
+## 2.2.0-beta.6
+
+### Patch Changes
+
+- Updated dependencies [cddfbf9]
+- Updated dependencies [a6e5f89]
+  - @osdk/api@2.2.0-beta.6
+
+## 2.2.0-beta.5
+
+### Patch Changes
+
+- @osdk/api@2.2.0-beta.5
+
+## 2.2.0-beta.4
+
+### Patch Changes
+
+- @osdk/api@2.2.0-beta.4
+
+## 2.2.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [761e69d]
+  - @osdk/api@2.2.0-beta.3
+
+## 2.2.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [fccd9ac]
+  - @osdk/api@2.2.0-beta.2
+
 ## 2.1.0-beta.31
 
 ### Patch Changes